refactor: split into multi-module monorepo with core/ shared source
* Root pom becomes packaging=pom aggregator (jib-jvm-flags-extension-parent).
* maven/ becomes the only Maven module under the aggregator. groupId, artifactId,
  version, JvmFlagsExtension FQN, and SPI service file path are all preserved —
  existing users upgrade with no pom changes.
* JvmFlagsLayerPlan and its test move from `jib.maven` to `jib.core` and live
  under core/src/{main,test}/java/...; core/ is shared source dir, not a Maven
  module (per ADR-0001). maven module pulls them in via build-helper-maven-plugin
  (add-source / add-test-source). Outlier note: the package move is a source-
  incompat tweak for direct importers of the helper class — to be called out in
  release notes when 1.1.0 ships.
* Drop license-maven-plugin; use Spotless's licenseHeader step instead, with
  the template updated to spotless `$YEAR` syntax.
* Declare lombok annotation processor path explicitly in maven-compiler-plugin
  so builds pass on JDK 23+ (per spike R4 finding).
* Pin google-java-format to 1.17.0 (last JDK-11-compatible release).

Implements PR-A from docs/specs/2026-05-monorepo-and-gradle-plan.md.
Refs #69
